autocomplete + geocode address->lat/lng->dark sky weather result
There are a few issues, mostly with presentation. As of right now, it seems that h:inputText cannot be hidden, and the versions that CAN be hidden (input or h:inputHidden) do not properly bind the bean to the searchedWeatherController...ie the lat and lng won't register. So I have ended up just disabling the fields, but they are still there.